Copy freestanding initrds in image mode
Copy freestand initrd files when cbm is run in image mode. Note the
initrd installation call happens prior to kernel install. This is done
as failure to install/update an initrd is fatal and if kernel install
happens before initrd install then an installed kernel could exist
with a configuration file referencing initrds that won't be available.